server/requirements.txt
Django==3.2.8
channels==3.0.4
django-cors-headers==3.8.0
django-shortuuidfield==0.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
#Pillow==8.3.2
pillow==10.2.0
shortuuid==1.0.1
psycopg2-binary==2.9.6
drf-spectacular==0.24.2
